무기에 따른 이펙트 변경 처리 완료, 방패 들기 동작 완료

+++ Client/CCollisionMgr.h
@@ -20,7 +20,7 @@ class CCollisionMgr
 	SINGLE(CCollisionMgr);
 
 private:
-	WORD m_Matrix[(UINT)ELAYER::END];
+	UINT m_Matrix[(UINT)ELAYER::END];
 	map<UINT_PTR, bool>	m_mapPrevInfo;	// 이전 프레임의 충돌여부
 
 public:
